move utils.load_data to io.load_data
ref: https://github.com/ME-ICA/tedana/issues/135
