+++ tedana/tests/test_utils.py
@@ -8,7 +8,7 @@ import nibabel as nib
 import numpy as np
 import pytest
 
-from tedana import utils
+from tedana import (utils, io)
 
 rs = np.random.RandomState(1234)
 datadir = pjoin(dirname(__file__), 'data')
@@ -118,35 +118,9 @@ def test_load_image():
     assert utils.load_image(fimg.get_data()).shape == exp_shape
 
 
-def test_load_data():
-    fimg = [nib.load(f) for f in fnames]
-    exp_shape = (64350, 3, 5)
-
-    # list of filepath to images
-    d, ref = utils.load_data(fnames, n_echos=len(tes))
-    assert d.shape == exp_shape
-    assert isinstance(ref, nib.Nifti1Image)
-    assert np.allclose(ref.get_data(), nib.load(fnames[0]).get_data())
-
-    # list of img_like
-    d, ref = utils.load_data(fimg, n_echos=len(tes))
-    assert d.shape == exp_shape
-    assert isinstance(ref, nib.Nifti1Image)
-    assert ref == fimg[0]
-
-    # imagine z-cat img
-    d, ref = utils.load_data(fnames[0], n_echos=3)
-    assert d.shape == (21450, 3, 5)
-    assert isinstance(ref, nib.Nifti1Image)
-    assert ref.shape == (39, 50, 11)
-
-    with pytest.raises(ValueError):
-        utils.load_data(fnames[0])
-
-
 def test_make_adaptive_mask():
     # load data make masks
-    data = utils.load_data(fnames, n_echos=len(tes))[0]
+    data = io.load_data(fnames, n_echos=len(tes))[0]
     minmask = utils.make_adaptive_mask(data)
     mask, masksum = utils.make_adaptive_mask(data, minimum=False, getsum=True)
 
@@ -174,7 +148,7 @@ def test_make_adaptive_mask():
 
 def test_make_min_mask():
     # load data make mask
-    data = utils.load_data(fnames, n_echos=len(tes))[0]
+    data = io.load_data(fnames, n_echos=len(tes))[0]
     minmask = utils.make_min_mask(data)
 
     assert minmask.shape == (64350,)
