+++ tedana/workflows/t2smap.py
@@ -154,7 +154,7 @@ def t2smap_workflow(data, tes, mask=None, fitmode='all', combmode='t2s',
         data = [data]
 
     LGR.info('Loading input data: {}'.format([f for f in data]))
-    catd, ref_img = utils.load_data(data, n_echos=n_echos)
+    catd, ref_img = io.load_data(data, n_echos=n_echos)
     n_samp, n_echos, n_vols = catd.shape
     LGR.debug('Resulting data shape: {}'.format(catd.shape))
